Updates from PR review: Relocate EmailProviderBase; Fix alphabetization in csproj; Remove unnecessary newlines
