+++ src/AndcultureCode.CSharp.Core/AndcultureCode.CSharp.Core.csproj
@@ -19,11 +19,11 @@
 
   <ItemGroup>
     <PackageReference Include="AndcultureCode.CSharp.Extensions" Version="0.2.0" />
+    <PackageReference Include="MailKit" Version="2.3.1.6" />
     <PackageReference Include="Microsoft.AspNetCore" Version="2.2.0" />
     <PackageReference Include="Microsoft.AspNetCore.Hosting.Abstractions" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.Localization.Abstractions" Version="3.1.0" />
-    <PackageReference Include="MailKit" Version="2.3.1.6" />
     <PackageReference Include="Vsxmd" Version="1.4.5">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
